Rename api/config_entry_ingress.go to api/config_entry_gateways.go
- Document the new ingress structs
